library_management_system/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
bcrypt==4.0.1
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==44.0.0
ecdsa==0.19.0
exceptiongroup==1.2.2
fastapi==0.115.5
greenlet==3.1.1
h11==0.14.0
idna==3.10
jwt==1.3.1
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.2
pydantic_core==2.27.1
PyMySQL==1.1.1
python-jose==3.3.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.32.1
